requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==24.1.0
ffmpeg-python==0.2.0
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
pandas==2.2.2
tqdm==4.66.4
transformers==4.43.3
huggingface-hub==0.23.4
pyannote.audio==3.3.2
torchmetrics==1.4.0
lightning==2.4.0
numba==0.60.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

